chore: apply pr review comments
