+++ web-client/src/styles/globals.css
@@ -34,7 +34,7 @@
     --chart-5: 27 87% 67%;
     --radius: 0.5rem;
     --sidebar-background: 0 0% 98%;
-    /* --sidebar-foreground: 240 5.3% 26.1%; */
+    --sidebar-foreground: 0 0% 0%;
     --sidebar-primary: 240 5.9% 10%;
     --sidebar-primary-foreground: 0 0% 98%;
     --sidebar-accent: 240 4.8% 95.9%;
